rename feedforward to mlp
