+++ slm_lab/agent/net/__init__.py
@@ -4,6 +4,6 @@ Contains classes of neural network architectures
 '''
 
 from slm_lab.agent.net.convnet import *
-from slm_lab.agent.net.feedforward import *
+from slm_lab.agent.net.mlp import *
 from slm_lab.agent.net.mlp_policy import *
 from slm_lab.agent.net.recurrent import *
